+++ tests/assets/c_source_2/header_with_types.h
@@ -1,0 +1,4 @@
+/**
+ * This should always be visible, even if ``no-private-members`` is in use.
+ */
+typedef float header_type;